// src/lib/types.ts
export interface Track {
	id: number;
	title: string;
	duration: number;
	replayGain?: number;
	peak?: number;
	allowStreaming: boolean;
	streamReady: boolean;
	streamStartDate?: string;
	premiumStreamingOnly: boolean;
	trackNumber: number;
	volumeNumber: number;
	version: string | null;
	popularity: number;
	copyright?: string;
	url: string;
	isrc?: string;
	editable: boolean;
	explicit: boolean;
	audioQuality: string;
	audioModes: string[];
	artist: Artist;
	artists: Artist[];
	album: Album;
	mixes?: Record<string, string>;
	mediaMetadata?: {
		tags: string[];
	};
}

export interface Artist {
	id: number;
	name: string;
	type: string;
	picture?: string;
	url?: string;
	popularity?: number;
	artistTypes?: string[];
	artistRoles?: Array<{
		category: string;
		categoryId: number;
	}>;
	mixes?: Record<string, string>;
}

export interface ArtistDetails extends Artist {
	albums: Album[];
	tracks: Track[];
}

export interface Album {
	id: number;
	title: string;
	cover: string;
	videoCover: string | null;
	releaseDate?: string;
	duration?: number;
	numberOfTracks?: number;
	numberOfVideos?: number;
	numberOfVolumes?: number;
	explicit?: boolean;
	popularity?: number;
	type?: string;
	upc?: string;
	copyright?: string;
	artist?: Artist;
	artists?: Artist[];
	audioQuality?: string;
	audioModes?: string[];
	url?: string;
	vibrantColor?: string;
	streamReady?: boolean;
	allowStreaming?: boolean;
	mediaMetadata?: {
		tags: string[];
	};
}